Fixes #158. Only compile C09Adaptor.cc under C09 so that the linker doesn't complain that C09Adaptor.o has no symbols.

+++ src/orc/C09Adapter.hh
@@ -42,7 +42,7 @@
 
   namespace std {
     // A poor man's stoll that converts str to a long long int base 10
-    long long stoll(std::string str);
+    int64_t stoll(std::string str);
   } // namespace std
